requirements.txt
numpy==1.26.4
pandas
pyrootutils==1.0.4
pyyaml==6.0.1
ultralytics
jupyter
opencv-python
onnx
# onnxruntime # CPU
onnxruntime-gpu # CUDA
openvino==2024.1.0
kornia
torch
torchvision